requirements.txt
Flask
flask-sock
paho-mqtt
python-dotenv
flasgger
